Add app file
